Apply review comment in PR
