+++ src/webapp/pages/form-page/disease-outbreak-event/utils/mapEntityToInitialFormState.ts
@@ -1,22 +1,9 @@
-import {
-    DiseaseOutbreakEventBaseAttrs,
-    IncidentStatusType,
-} from "../../../../../domain/entities/disease-outbreak-event/DiseaseOutbreakEvent";
 import { DiseaseOutbreakEventWithOptions } from "../../../../../domain/entities/disease-outbreak-event/DiseaseOutbreakEventWithOptions";
-import {
-    FormFieldState,
-    FormState,
-    getAllFieldsFromSections,
-    getDateFieldValue,
-    getMultipleOptionsFieldValue,
-    getBooleanFieldValue,
-    getStringFieldValue,
-} from "../../../../components/form/FormState";
 import { Option } from "../../../../../domain/entities/Ref";
-import { getFieldIdFromIdsDictionary } from "../../../../components/form/FormState";
+import { getFieldIdFromIdsDictionary } from "../../../../components/form/FormFieldsState";
+import { FormState } from "../../../../components/form/FormState";
 import { UserOption } from "../../../../components/user-selector/UserSelector";
 import { Option as PresentationOption } from "../../../../components/utils/option";
-import { getHazardTypeValue } from "../../../../../data/repositories/consts/DiseaseOutbreakConstants";
 
 export const diseaseOutbreakEventFieldIds = {
     name: "name",
@@ -49,6 +36,7 @@ export const diseaseOutbreakEventFieldIds = {
     notes: "notes",
 } as const;
 
+// TODO: Thinking for the future about generate this FormState by iterating over Object.Keys(diseaseOutbreakEvent)
 export function mapEntityToInitialFormState(
     diseaseOutbreakEventWithOptions: DiseaseOutbreakEventWithOptions
 ): FormState {
@@ -668,139 +656,4 @@ function mapToPresentationOptions(options: Option[]): PresentationOption[] {
             label: option.name,
         })
     );
-}
-
-export function mapFormStateToEntityData(
-    formState: FormState,
-    currentUserName: string,
-    diseaseOutbreakEventWithOptions: DiseaseOutbreakEventWithOptions
-): DiseaseOutbreakEventBaseAttrs {
-    const { diseaseOutbreakEvent } = diseaseOutbreakEventWithOptions;
-
-    const allFields: FormFieldState[] = getAllFieldsFromSections(formState.sections);
-
-    const hazardType = getStringFieldValue(diseaseOutbreakEventFieldIds.hazardType, allFields);
-
-    const diseaseOutbreakEventEditableData = {
-        name: getStringFieldValue(diseaseOutbreakEventFieldIds.name, allFields),
-        hazardType: getHazardTypeValue(hazardType),
-        mainSyndromeCode: getStringFieldValue(
-            diseaseOutbreakEventFieldIds.mainSyndromeCode,
-            allFields
-        ),
-        suspectedDiseaseCode: getStringFieldValue(
-            diseaseOutbreakEventFieldIds.suspectedDiseaseCode,
-            allFields
-        ),
-        notificationSourceCode: getStringFieldValue(
-            diseaseOutbreakEventFieldIds.notificationSourceCode,
-            allFields
-        ),
-        areasAffectedProvinceIds: getMultipleOptionsFieldValue(
-            diseaseOutbreakEventFieldIds.areasAffectedProvinceIds,
-            allFields
-        ),
-        areasAffectedDistrictIds: getMultipleOptionsFieldValue(
-            diseaseOutbreakEventFieldIds.areasAffectedDistrictIds,
-            allFields
-        ),
-        incidentStatus: getStringFieldValue(
-            diseaseOutbreakEventFieldIds.incidentStatus,
-            allFields
-        ) as IncidentStatusType,
-        emerged: {
-            date: getDateFieldValue(diseaseOutbreakEventFieldIds.emergedDate, allFields) as Date,
-            narrative: getStringFieldValue(
-                diseaseOutbreakEventFieldIds.emergedNarrative,
-                allFields
-            ),
-        },
-        detected: {
-            date: getDateFieldValue(diseaseOutbreakEventFieldIds.detectedDate, allFields) as Date,
-            narrative: getStringFieldValue(
-                diseaseOutbreakEventFieldIds.detectedNarrative,
-                allFields
-            ),
-        },
-        notified: {
-            date: getDateFieldValue(diseaseOutbreakEventFieldIds.notifiedDate, allFields) as Date,
-            narrative: getStringFieldValue(
-                diseaseOutbreakEventFieldIds.notifiedNarrative,
-                allFields
-            ),
-        },
-        earlyResponseActions: {
-            initiateInvestigation: getDateFieldValue(
-                diseaseOutbreakEventFieldIds.initiateInvestigation,
-                allFields
-            ) as Date,
-            conductEpidemiologicalAnalysis: getDateFieldValue(
-                diseaseOutbreakEventFieldIds.conductEpidemiologicalAnalysis,
-                allFields
-            ) as Date,
-            laboratoryConfirmation: {
-                date: getDateFieldValue(
-                    diseaseOutbreakEventFieldIds.laboratoryConfirmationDate,
-                    allFields
-                ) as Date,
-                na: getBooleanFieldValue(
-                    diseaseOutbreakEventFieldIds.laboratoryConfirmationNA,
-                    allFields
-                ),
-            },
-            appropriateCaseManagement: {
-                date: getDateFieldValue(
-                    diseaseOutbreakEventFieldIds.appropriateCaseManagementDate,
-                    allFields
-                ) as Date,
-                na: getBooleanFieldValue(
-                    diseaseOutbreakEventFieldIds.appropriateCaseManagementNA,
-                    allFields
-                ),
-            },
-            initiatePublicHealthCounterMeasures: {
-                date: getDateFieldValue(
-                    diseaseOutbreakEventFieldIds.initiatePublicHealthCounterMeasuresDate,
-                    allFields
-                ) as Date,
-                na: getBooleanFieldValue(
-                    diseaseOutbreakEventFieldIds.initiatePublicHealthCounterMeasuresNA,
-                    allFields
-                ),
-            },
-            initiateRiskCommunication: {
-                date: getDateFieldValue(
-                    diseaseOutbreakEventFieldIds.initiateRiskCommunicationDate,
-                    allFields
-                ) as Date,
-                na: getBooleanFieldValue(
-                    diseaseOutbreakEventFieldIds.initiateRiskCommunicationNA,
-                    allFields
-                ),
-            },
-            establishCoordination: getDateFieldValue(
-                diseaseOutbreakEventFieldIds.establishCoordination,
-                allFields
-            ) as Date,
-            responseNarrative: getStringFieldValue(
-                diseaseOutbreakEventFieldIds.responseNarrative,
-                allFields
-            ),
-        },
-        incidentManagerName: getStringFieldValue(
-            diseaseOutbreakEventFieldIds.incidentManagerName,
-            allFields
-        ),
-        notes: getStringFieldValue(diseaseOutbreakEventFieldIds.notes, allFields),
-    };
-
-    const diseaseOutbreakEventBase: DiseaseOutbreakEventBaseAttrs = {
-        id: diseaseOutbreakEvent?.id || "",
-        created: diseaseOutbreakEvent?.created || new Date(),
-        lastUpdated: diseaseOutbreakEvent?.lastUpdated || new Date(),
-        createdByName: diseaseOutbreakEvent?.createdByName || currentUserName,
-        ...diseaseOutbreakEventEditableData,
-    };
-
-    return diseaseOutbreakEventBase;
 }
