+++ src/webapp/pages/form-page/disease-outbreak-event/utils/mapFormStateToEntityData.ts
@@ -1,0 +1,151 @@
+import {
+    DiseaseOutbreakEventBaseAttrs,
+    IncidentStatusType,
+} from "../../../../../domain/entities/disease-outbreak-event/DiseaseOutbreakEvent";
+import { DiseaseOutbreakEventWithOptions } from "../../../../../domain/entities/disease-outbreak-event/DiseaseOutbreakEventWithOptions";
+import { FormState } from "../../../../components/form/FormState";
+import { getHazardTypeValue } from "../../../../../data/repositories/consts/DiseaseOutbreakConstants";
+import { diseaseOutbreakEventFieldIds } from "./mapEntityToInitialFormState";
+import {
+    FormFieldState,
+    getAllFieldsFromSections,
+    getBooleanFieldValue,
+    getDateFieldValue,
+    getMultipleOptionsFieldValue,
+    getStringFieldValue,
+} from "../../../../components/form/FormFieldsState";
+
+export function mapFormStateToEntityData(
+    formState: FormState,
+    currentUserName: string,
+    diseaseOutbreakEventWithOptions: DiseaseOutbreakEventWithOptions
+): DiseaseOutbreakEventBaseAttrs {
+    const { diseaseOutbreakEvent } = diseaseOutbreakEventWithOptions;
+
+    const allFields: FormFieldState[] = getAllFieldsFromSections(formState.sections);
+
+    const hazardType = getStringFieldValue(diseaseOutbreakEventFieldIds.hazardType, allFields);
+
+    const diseaseOutbreakEventEditableData = {
+        name: getStringFieldValue(diseaseOutbreakEventFieldIds.name, allFields),
+        hazardType: getHazardTypeValue(hazardType),
+        mainSyndromeCode: getStringFieldValue(
+            diseaseOutbreakEventFieldIds.mainSyndromeCode,
+            allFields
+        ),
+        suspectedDiseaseCode: getStringFieldValue(
+            diseaseOutbreakEventFieldIds.suspectedDiseaseCode,
+            allFields
+        ),
+        notificationSourceCode: getStringFieldValue(
+            diseaseOutbreakEventFieldIds.notificationSourceCode,
+            allFields
+        ),
+        areasAffectedProvinceIds: getMultipleOptionsFieldValue(
+            diseaseOutbreakEventFieldIds.areasAffectedProvinceIds,
+            allFields
+        ),
+        areasAffectedDistrictIds: getMultipleOptionsFieldValue(
+            diseaseOutbreakEventFieldIds.areasAffectedDistrictIds,
+            allFields
+        ),
+        incidentStatus: getStringFieldValue(
+            diseaseOutbreakEventFieldIds.incidentStatus,
+            allFields
+        ) as IncidentStatusType,
+        emerged: {
+            date: getDateFieldValue(diseaseOutbreakEventFieldIds.emergedDate, allFields) as Date,
+            narrative: getStringFieldValue(
+                diseaseOutbreakEventFieldIds.emergedNarrative,
+                allFields
+            ),
+        },
+        detected: {
+            date: getDateFieldValue(diseaseOutbreakEventFieldIds.detectedDate, allFields) as Date,
+            narrative: getStringFieldValue(
+                diseaseOutbreakEventFieldIds.detectedNarrative,
+                allFields
+            ),
+        },
+        notified: {
+            date: getDateFieldValue(diseaseOutbreakEventFieldIds.notifiedDate, allFields) as Date,
+            narrative: getStringFieldValue(
+                diseaseOutbreakEventFieldIds.notifiedNarrative,
+                allFields
+            ),
+        },
+        earlyResponseActions: {
+            initiateInvestigation: getDateFieldValue(
+                diseaseOutbreakEventFieldIds.initiateInvestigation,
+                allFields
+            ) as Date,
+            conductEpidemiologicalAnalysis: getDateFieldValue(
+                diseaseOutbreakEventFieldIds.conductEpidemiologicalAnalysis,
+                allFields
+            ) as Date,
+            laboratoryConfirmation: {
+                date: getDateFieldValue(
+                    diseaseOutbreakEventFieldIds.laboratoryConfirmationDate,
+                    allFields
+                ) as Date,
+                na: getBooleanFieldValue(
+                    diseaseOutbreakEventFieldIds.laboratoryConfirmationNA,
+                    allFields
+                ),
+            },
+            appropriateCaseManagement: {
+                date: getDateFieldValue(
+                    diseaseOutbreakEventFieldIds.appropriateCaseManagementDate,
+                    allFields
+                ) as Date,
+                na: getBooleanFieldValue(
+                    diseaseOutbreakEventFieldIds.appropriateCaseManagementNA,
+                    allFields
+                ),
+            },
+            initiatePublicHealthCounterMeasures: {
+                date: getDateFieldValue(
+                    diseaseOutbreakEventFieldIds.initiatePublicHealthCounterMeasuresDate,
+                    allFields
+                ) as Date,
+                na: getBooleanFieldValue(
+                    diseaseOutbreakEventFieldIds.initiatePublicHealthCounterMeasuresNA,
+                    allFields
+                ),
+            },
+            initiateRiskCommunication: {
+                date: getDateFieldValue(
+                    diseaseOutbreakEventFieldIds.initiateRiskCommunicationDate,
+                    allFields
+                ) as Date,
+                na: getBooleanFieldValue(
+                    diseaseOutbreakEventFieldIds.initiateRiskCommunicationNA,
+                    allFields
+                ),
+            },
+            establishCoordination: getDateFieldValue(
+                diseaseOutbreakEventFieldIds.establishCoordination,
+                allFields
+            ) as Date,
+            responseNarrative: getStringFieldValue(
+                diseaseOutbreakEventFieldIds.responseNarrative,
+                allFields
+            ),
+        },
+        incidentManagerName: getStringFieldValue(
+            diseaseOutbreakEventFieldIds.incidentManagerName,
+            allFields
+        ),
+        notes: getStringFieldValue(diseaseOutbreakEventFieldIds.notes, allFields),
+    };
+
+    const diseaseOutbreakEventBase: DiseaseOutbreakEventBaseAttrs = {
+        id: diseaseOutbreakEvent?.id || "",
+        created: diseaseOutbreakEvent?.created || new Date(),
+        lastUpdated: diseaseOutbreakEvent?.lastUpdated || new Date(),
+        createdByName: diseaseOutbreakEvent?.createdByName || currentUserName,
+        ...diseaseOutbreakEventEditableData,
+    };
+
+    return diseaseOutbreakEventBase;
+}
